Proiect PAO 2 Mai Varianta 1

+++ src/proiect/daoservices/StadionRepositoryService.java
@@ -1,18 +1,48 @@
 package proiect.daoservices;
 
+import proiect.model.Echipa;
 import proiect.model.Stadion;
 import proiect.dao.StadionDao;
+import proiect.utils.InvalidDataException;
 
+import java.sql.SQLException;
 import java.util.List;
 
 public class StadionRepositoryService {
-    private final StadionDao stadionDao = new StadionDao();
+    private StadionDao stadionDao = StadionDao.getInstance();
 
-    public void addStadion(Stadion stadion) {
-        stadionDao.create(stadion);
+    public StadionRepositoryService() throws SQLException {}
+
+    public void printAllStadiums() throws SQLException {
+        List<Stadion> categories = stadionDao.findAllStadion();
+        if(categories != null){
+            categories.forEach(System.out:: println);
+        } else {
+            System.out.println("There is no category.");
+        }
     }
 
-    public Stadion getStadionByName(String name) {
+    public List<Stadion> getAllStadions() throws SQLException {
+        return stadionDao.findAllStadion();
+    }
+
+    public void addStadion(Stadion stadion) throws InvalidDataException {
+        try {
+            if(stadion != null){
+                if(stadionDao.read(String.valueOf(stadion.getId())) != null)
+                    throw new InvalidDataException("We already have a stadium with this ID!");
+                if(stadionDao.read(stadion.getNume()) != null)
+                    throw new InvalidDataException("We already have a stadium with this name!");
+
+                stadionDao.create(stadion);
+                System.out.println("Stadium added!");
+            }
+        } catch (SQLException e) {
+            System.out.println("Creation failed: " + e.getMessage());
+        }
+    }
+
+    public Stadion getStadionByName(String name) throws SQLException {
         Stadion stadion = stadionDao.read(name);
 
         if (stadion != null) {
@@ -24,15 +54,47 @@ public class StadionRepositoryService {
         return null;
     }
 
-    public void updateStadion(String name, Stadion updatedStadion) {
-        stadionDao.update(name, updatedStadion);
+    public Stadion getStadionById(int id) throws SQLException {
+        Stadion stadion = stadionDao.readByID(id);
+
+        if (stadion != null) {
+            return stadion;
+        } else {
+            System.out.println("Stadium has not been found!");
+        }
+
+        return null;
+    }
+
+    public void updateStadion(String name, Stadion updatedStadion) throws InvalidDataException {
+        try {
+            Stadion stadion = stadionDao.read(name);
+            if (stadion != null) {
+                if(!stadionDao.checkUniqueName(stadion.getNume()))
+                    throw new InvalidDataException("We already have a stadium with this name!");
+            }
+
+            stadionDao.update(name, updatedStadion);
+            System.out.println("Stadium updated!");
+        } catch (SQLException e) {
+            System.out.println("Update failed: " + e.getMessage());
+        }
     }
 
     public void removeStadion(Stadion stadion) {
-        stadionDao.delete(stadion);
-    }
+        try {
+            if (stadion == null) return;
 
-    public List<Stadion> getAllStadions() {
-        return stadionDao.findAllStadion();
+            List<Echipa> echipe = stadion.getEchipe();
+            if(echipe != null) {
+                for(Echipa e : echipe) {
+                    e.setStadion(null);
+                }
+            }
+            stadionDao.delete(stadion);
+
+        } catch (SQLException e) {
+            System.out.println("SQLException " + e.getSQLState() + " " + e.getMessage());
+        }
     }
 }
